threaded_watch_cam/requirements.txt
flask
numpy
opencv-contrib-python
imutils
pydrive
python-dotenv
gunicorn
gevent